Add support for multiple user profiles
This adds support for multiple "user" profile queries by specifying
their configuration through additional environment variables that suffix
a label to the standard ones.

For example, *in addition to the standard variables*, one could specify
the same set of variables with "_FOO" suffixed to the environment
variable name: FINGER_NAME_FOO, FINGER_PROJECT_FOO, and so on, where "_"
is a required separator and "FOO" is a unique label for the set of
variables for this user profile configuration.

This is backwards compatible and the default set of variables are still
required.